+++ src/main/kotlin/com/github/asoee/cursorlessjetbrains/settings/TalonSettingsConfigurable.kt
@@ -2,6 +2,8 @@ package com.github.asoee.cursorlessjetbrains.settings
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.options.Configurable
+import com.intellij.openapi.options.Configurable.isFieldModified
+import com.jetbrains.rd.util.reflection.usingTrueFlag
 import org.jetbrains.annotations.Nls
 import org.jetbrains.annotations.Nullable
 import java.util.*
@@ -30,7 +32,8 @@ internal class TalonSettingsConfigurable : Configurable {
             Objects.requireNonNull(TalonSettings.instance.state)
         return mySettingsComponent!!.hatsScaleFactor.toInt() != state.hatScaleFactor ||
                 mySettingsComponent!!.hatsVerticalOffset.toInt() != state.hatVerticalOffset ||
-                mySettingsComponent!!.enableHats != state.enableHats
+                mySettingsComponent!!.enableHats != state.enableHats ||
+                mySettingsComponent!!.flasRangeDuration.toInt() != state.flashRangeDuration
     }
 
     override fun apply() {
@@ -39,6 +42,7 @@ internal class TalonSettingsConfigurable : Configurable {
         state.hatScaleFactor = mySettingsComponent!!.hatsScaleFactor.toInt()
         state.hatVerticalOffset = mySettingsComponent!!.hatsVerticalOffset.toInt()
         state.enableHats = mySettingsComponent!!.enableHats
+        state.flashRangeDuration = mySettingsComponent!!.flasRangeDuration.toInt()
 
         val messageBus = ApplicationManager.getApplication().messageBus
         messageBus.syncPublisher(TalonSettingsListener.TOPIC).onSettingsChanged(state)
@@ -51,6 +55,7 @@ internal class TalonSettingsConfigurable : Configurable {
         mySettingsComponent!!.hatsScaleFactor = state.hatScaleFactor.toString()
         mySettingsComponent!!.hatsVerticalOffset = state.hatVerticalOffset.toString()
         mySettingsComponent!!.enableHats = state.enableHats
+        mySettingsComponent!!.flasRangeDuration = state.flashRangeDuration.toString()
     }
 
     override fun disposeUIResources() {