+++ src/main/kotlin/com/github/asoee/cursorlessjetbrains/vscode/VsCodeSettingsService.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.Disposable
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.util.SystemInfo
 import com.intellij.util.EventDispatcher
+import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
 import java.nio.file.Path
@@ -12,6 +13,7 @@ import kotlin.io.path.Path
 import kotlin.io.path.absolutePathString
 import kotlin.io.path.exists
 
+@OptIn(ExperimentalSerializationApi::class)
 @Service(Service.Level.APP)
 class VsCodeSettingsService() : Disposable {
 
@@ -86,12 +88,14 @@ class VsCodeSettingsService() : Disposable {
         this.watcher?.dispose()
     }
 
+    private val json = Json { allowTrailingComma = true }
+
     fun currentSettings(): VsCodeSettings? {
         val fullPath = this.vsCodeSettingsDir?.resolve("settings.json")
         if (fullPath != null && fullPath.exists()) {
             val jsonString = fullPath.toFile().readText()
             println("FileCommandServer: settings.json content: $jsonString")
-            val jsonObject = Json.decodeFromString<JsonObject>(jsonString)
+            val jsonObject = json.decodeFromString<JsonObject>(jsonString)
             val settings = VsCodeSettings(jsonObject)
             return settings
         }